app/requirements.txt
flask
python-dotenv
oracledb
gunicorn
